refactor(common): split lib/common.js into smaller files to improve maintainability
Prepare for adding lib/TypeScript to support TypeScript
