+++ lib/ConfitGenerator.js
@@ -17,9 +17,13 @@ Base.prototype._getStorage = function () {
 
 
 // Extend the Base prototype
+_.extend(Base.prototype, require('./buildToolUtils.js'));
 _.extend(Base.prototype, require('./fileUtils.js'));
+_.extend(Base.prototype, require('./npm.js'));
+_.extend(Base.prototype, require('./resources.js'));
 _.extend(Base.prototype, require('./utils.js'));
 
+//console.log(Base.prototype);
 
 module.exports = {
